+++ examples/opensora_hpcai/tools/caption/PLLaVA/models/llama/network.py
@@ -31,7 +31,7 @@ class LlamaDecoderLayer(nn.Cell):
         attention_dropout: float = 0.0,
         hidden_act: str = "silu",
         attn_implementation: Literal["eager", "flash_attention"] = "eager",
-        dtype: ms.dtype = ms.float32,
+        dtype: ms.dtype = ms.bfloat16,
     ) -> None:
         super().__init__()
         self.hidden_size = hidden_size
@@ -101,7 +101,7 @@ class LlamaModel(nn.Cell):
         hidden_act: str = "silu",
         pad_token_id: Optional[int] = None,
         attn_implementation: Literal["eager", "flash_attention"] = "eager",
-        dtype: ms.dtype = ms.float32,
+        dtype: ms.dtype = ms.bfloat16,
     ) -> None:
         super().__init__()
         self.padding_idx = pad_token_id
@@ -232,7 +232,7 @@ class LlamaForCausalLM(nn.Cell):
         hidden_act: str = "silu",
         pad_token_id: Optional[int] = None,
         attn_implementation: Literal["eager", "flash_attention"] = "eager",
-        dtype: ms.dtype = ms.float32,
+        dtype: ms.dtype = ms.bfloat16,
         **kwargs,
     ) -> None:
         super().__init__()
@@ -295,5 +295,5 @@ class LlamaForCausalLM(nn.Cell):
             past_value_cache_list=past_value_cache_list,
             return_key_value_cache=return_key_value_cache,
         )
-        logits = self.lm_head(hidden_states).to(ms.float32)
+        logits = self.lm_head(hidden_states).to(ms.bfloat16)
         return logits, key_cache_list, value_cache_list
